(Qt) Combine shaderparamsdialog.cpp and qt_dialogs.cpp - further binary size reduction
