+++ griffin/griffin_cpp.cpp
@@ -53,9 +53,6 @@ UI
 
 #include "../ui/drivers/qt/ui_qt_load_core_window.cpp"
 #include "../ui/drivers/qt/gridview.cpp"
-#if defined(HAVE_CG) || defined(HAVE_GLSL) || defined(HAVE_SLANG) || defined(HAVE_HLSL)
-#include "../ui/drivers/qt/shaderparamsdialog.cpp"
-#endif
 #include "../ui/drivers/qt/qt_dialogs.cpp"
 #include "../ui/drivers/qt/filedropwidget.cpp"
 #include "../ui/drivers/qt/qt_playlist.cpp"
